Change type int to uint32_t when it's necessary
I also removed unused variables

+++ src/common/view_port.cpp
@@ -115,12 +115,12 @@ box_t viewport_t::frame_to_screen_coords(uint32_t video_w, uint32_t video_h, con
 }
 
 float viewport_t::fit(uint32_t *w, uint32_t *h) const {
-    if (*w <= vp_[2] && *h <= vp_[3]) {
+    if (static_cast<int>(*w) <= vp_[2] && static_cast<int>(*h) <= vp_[3]) {
         return 1.0;  // no upscale
     }
 
-    float fx = static_cast<float>(vp_[2]);
-    float fy = static_cast<float>(vp_[3]);
+    float fx = vp_[2];
+    float fy = vp_[3];
 
     float xscale = fx / *w;
     float scale = fy / *h;
